Add a webserver (McWeb) to handle the Minecraft server
